+++ src/main/java/com/cherries/exchange/demo/entities/User.java
@@ -15,7 +15,7 @@ public class User {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name="id")
-    private BigInteger id;
+    private Integer id;
 
     @Column(name="first_name")
     private String first_name;
